Refactor install.sh so chsh happen last
This is so the script doesn't bomb in a half done state.

Before, if sudo wasn't correctly entered for package manager
commands, the script doesn't do anything. This is good.

Before, if the `sudo chsh` failed, then the script is in a
half done state, which is bad.

Now, `sudo chsh` is at the end, so it doesn't matter if it
succeeds.